services/policyengine_svc/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-dotenv>=1.0
opencensus-ext-azure>=1.1
